Moved Mesh to separate file
