+++ src/Model.cpp
@@ -1,59 +1,5 @@
 #include "Model.hpp"
 
-/* --- MESH --- */
-/* Public Mesh functions */
-
-void Mesh::upload_mesh_data()
-{
-    glGenVertexArrays(1, &this->VAO);
-    glBindVertexArray(this->VAO);
-    glGenBuffers(1, &this->EBO);
-    glGenBuffers(4, this->VBO);
-
-    GLuint* indices = &this->indices[0];
-    GLfloat* vertices = &this->vertices[0];
-    GLfloat* normals = &this->normals[0];
-    GLfloat* tex_coords = &this->tex_coords[0];
-    GLfloat* tangents = &this->tangents[0];
-
-    /* Element array buffer */
-    glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, this->EBO);
-    glBufferData(GL_ELEMENT_ARRAY_BUFFER, index_count * sizeof(GLuint), indices, GL_STATIC_DRAW);
-
-    /* Vertex coordinates */
-    glBindBuffer(GL_ARRAY_BUFFER, this->VBO[0]);
-    glBufferData(GL_ARRAY_BUFFER, 3 * vertex_count * sizeof(GLfloat), vertices, GL_STATIC_DRAW);
-    glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 0, 0);
-    glEnableVertexAttribArray(0);
-
-    /* Normal vectors */
-    glBindBuffer(GL_ARRAY_BUFFER, this->VBO[1]);
-    glBufferData(GL_ARRAY_BUFFER, 3 * vertex_count * sizeof(GLfloat), normals, GL_STATIC_DRAW);
-    glVertexAttribPointer(1, 3, GL_FLOAT, GL_FALSE, 0, 0);
-    glEnableVertexAttribArray(1);
-
-    /* Texture coordinates */
-    glBindBuffer(GL_ARRAY_BUFFER, this->VBO[2]);
-    glBufferData(GL_ARRAY_BUFFER, 2 * vertex_count * sizeof(GLfloat), tex_coords, GL_STATIC_DRAW);
-    glVertexAttribPointer(2, 2, GL_FLOAT, GL_FALSE, 0, 0);
-    glEnableVertexAttribArray(2);
-
-    /* Tangent vectors */
-    glBindBuffer(GL_ARRAY_BUFFER, this->VBO[3]);
-    glBufferData(GL_ARRAY_BUFFER, 3 * vertex_count * sizeof(GLfloat), tangents, GL_STATIC_DRAW);
-    glVertexAttribPointer(3, 3, GL_FLOAT, GL_FALSE, 0, 0);
-    glEnableVertexAttribArray(3);
-
-    /* Unbind VAO */
-    glBindVertexArray(0);
-}
-
-GLuint Mesh::get_VAO()
-{
-    return VAO;
-}
-
-
 /* --- MODEL ---*/
 std::vector<Model*> Model::loaded_models, Model::loaded_flat_models;
 std::vector<Texture*> Model::loaded_textures;
