+++ src/Model.hpp
@@ -29,44 +29,7 @@
 #include "Light.hpp"
 #include "Error.hpp"
 #include "Animation_Path.hpp"
-
-const std::string DEFAULT_PATH = "res/models/default";
-const std::string DEFAULT_DIFFUSE = "default_diffuse.jpg";
-const std::string DEFAULT_NORMAL = "default_normal.jpg";
-
-
-enum texture_type {
-    DIFFUSE,
-    SPECULAR,
-    NORMAL
-};
-
-struct Texture {
-    GLuint id;
-    texture_type type;
-    aiString path;
-};
-
-
-class Mesh {
-public:
-    GLuint index_count, vertex_count;
-    GLfloat shininess = 80;
-    std::vector<GLuint> indices;
-    std::vector<GLfloat> vertices, normals, tex_coords, tangents;
-    Texture *diffuse_map, *specular_map, *normal_map;
-
-    Mesh() { };
-    ~Mesh() { };
-
-    /* Upload vertices, normals etc to the GPU */
-    void upload_mesh_data();
-    GLuint get_VAO();
-
-private:
-    GLuint VAO, EBO;
-    GLuint VBO[4]; // Vertices, normals, texCoords, tangents
-};
+#include "Mesh.hpp"
 
 
 class Model {
